requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap4==2023.1
Django==4.2.10
django-crispy-forms==2.1
django-humanize==0.1.2
django-tinymce==3.7.1
humanize==4.9.0
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
